+++ src/drytorch/trackers/logging.py
@@ -197,6 +197,13 @@ class BuiltinLogger(tracking.Tracker):
         return super().notify(event)
 
     @notify.register
+    def _(self, event: log_events.PauseExperimentEvent) -> None:
+        msg = 'Experiment: %(name)s - Pausing run: %(id)s'
+        args = {'name': event.exp_name, 'id': event.run_id}
+        logger.log(INFO_LEVELS.experiment, msg, args)
+        return super().notify(event)
+
+    @notify.register
     def _(self, event: log_events.LearningRateEvent) -> None:
         message_parts = [
             'Updated %(model_name)s optimizer at epoch %(epoch)d',