+++ src/drytorch/trackers/wandb.py
@@ -131,6 +131,10 @@ class Wandb(Dumper):
         return super().notify(event)
 
     @notify.register
+    def _(self, event: log_events.PauseExperimentEvent) -> None:
+        return super().notify(event)
+
+    @notify.register
     def _(self, event: log_events.MetricEvent) -> None:
         """Process metric events.